requirements.txt
pika==1.3.2
requests==2.31.0
loguru==0.7.0
python-dotenv==1.0.0
SQLAlchemy==2.0.19

# поднимаем базу таким макаром
# docker run --rm -it --hostname my-rabbit -p 15672:15672 -p 5672:5672 rabbitmq:3-management
# отправляем
# python send.py
# получаем
# python receive.py